Use stricter DB API
> The optional parameters may be passed as a sequence, as specified by
> the DB API, or as individual values.

We were passing them as individual values, which this less strict API
allows. However, the `cursor` provided by SQLAlchemy adheres to the
stricter DB API, which requires them to be passed as sequences.

NOTE: These could be `[lists]` instead of `(tuples,)` which are a bit
nicer to read, especially for single values.

Similarly, this less strict API has a `commit()` method on the returned
cursor; however, the stricter DB API only has it on the `connection`
object. So again, this new method works for both interfaces.