Terrible fix for error in SQL query
The `in` had been dropped in a prior commit, and additionally the switch
from string formatting to qmark param for the `[revision] in ('foo',
'bar')` clause didn't work (probably due to the difficulty of getting
SQL queries right in Python). This is a quick fix before we delete this
code and replace it with ORM queries.