Remove `[dbo].` prefix
It's the default schema prefix. So while it is more explicit to specify
it, we already do so inconsistently, and the ORM schemas don't include
it. It will be looked up and used as necessary, so it's safe to
delete (minus a small performance penalty of that lookup, but I doubt
we're worried about this right now).