Fix scoping bug in MonitorDownstream.py
When I added the CLI I didn't realize that
`monitoring_subject_db_driver` was being used as a global to this file.
When the main script was turned into a function,
`update_tracked_revisions` lost access to the scope in which it lived.
So instead we just need to pass it.

While I had tested, since there were no revisions to update at the time,
this path wasn't executed. Also Python isn't statically typed so the
interpreter doesn't provide a warning. We should write some unit tests,
and add some static analysis to avoid such bugs in the future.