Rewrite the CLI to support adding a distro
It became apparent that we're going to have multiple subcommands. Also,
in order to get the CI to run within the timeout, we need to be able to
add just a single distro. So I rewrote the CLI to use two subcommands,
`run` and `add-distro`, which each take their own flags.

To facilitate this, I moved the default distros into `Util.Config` and
removed the responsibility of populating it from the `DatabaseDriver`.

By using the `argparse` module's subcommands feature, we can use the
default function to make executing these really clean.