Replace use of `os.system` with `Subprocess` module
This lets us avoid dealing with `cwd` issues because `subprocess.run()`
takes the directory as an argument. We can get the Git working directory
using the `repo` object. We can also use the `repo` object to handle the
Git reset commands, again avoiding `cwd` problems.

We also remove the need for the temporary file as the completed process
object has a `stdout` string we can just call `splitlines()` on.