Fix up clone/fetch to be more resilient and faster
We now clone and fetch only the last four years of history, which is
still a lot but significantly less than everything. The oldest missing
commits we're concerned with are less than three years of age, so this
should be sufficient, but perhaps should be configurable.

When fetching, we need to specify `--force` in case upstream tags are
changed (or collide with each other, which we don't really care about).
There's probably a better solution to this, but it works for now.

squash! Fix up clone/fetch to be more resilient and faster