Undo `--bare` cloning of linux.git
This just...causes problems. While we _don’t_ want a worktree, and
having one _may_ cause problems, a bare clone doesn’t maintain an
`origin/*` set of refs. Instead, it maps them into the local refs (that
is, what is normally `origin/master` is just `master`). In and of
itself, this is fine, but `git fetch --all` has problems with this, and
we do not want to fetch twice just to be able to update the local master
branch with the remote master branch. That is, `git fetch origin
master:master` does solve this problem, but requires a second fetch on
top of the `git fetch --all` we do because we have multiple remotes.
Finally, that local `master` branch, not being a remote namespace,
starts to get confused with the other remotes’ `master` branches. So we
are saying good-bye to the bare clone.